requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.10.0
django-audit-log==0.7.0
django-auditlog==0.4.3
django-crispy-forms==1.6.1
django-jsonfield==1.0.1
django-sslserver==0.19
packaging==16.8
pycparser==2.17
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
Celery==4.4.0
rabbitmq==0.2.0